pr(store): store разделен на newsListState и newsItemState
