+++ src/components/NewsContentBlock/NewsContentBlock.tsx
@@ -10,7 +10,7 @@ import {
   Time,
 } from './NewsContentBlock.styled.ts';
 import { NewsItemTitle } from '../../commons/NewsItemTitle/NewsItemTitle.ts';
-import { useSelectorNewsItem } from '../../store/store.ts';
+import { useSelectorNewsItem } from '../../store/states/newsItemState/newsItemState.ts';
 
 export default function NewsContentBlock() {
   const newsItem = useSelectorNewsItem();
