+++ src/store/states/newsItemState/newsItemState.types.ts
@@ -1,22 +1,4 @@
-export interface UseNewsListStateType {
-  newsList: NewsListType[] | [];
-  newsLoading: boolean;
-  newsServerDown: boolean;
-  getNewsList: () => Promise<void>;
-}
-
-export interface NewsListType {
-  id: number;
-  title: string;
-  points: number | null;
-  user: string | null;
-  time: number;
-  time_ago: string;
-  comments_count: number;
-  type: string;
-  url?: string;
-  domain?: string;
-}
+import { NewsListType } from '../newsListState/newsListState.types.ts';
 
 export interface UseNewsItemType {
   newsItem: NewsItemType | null;
